Refactor: Clean up SAINTH component and fix status command
- Removed unused state variable `audioLevel`.
- Removed unused refs: `animationRef`, `audioContextRef`, `analyserRef`.
- Updated the voice command for system status to report the correct number of connected devices using `devices.length` instead of a stale state variable.
